Move legacy write operations into driver-legacy
JAVA-4795
